requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
jinja2==3.1.2
# OCR and image processing
pytesseract==0.3.10
PyMuPDF==1.23.8
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.0
# LLM integration
google-generativeai==0.3.1
# Additional utilities
rapidfuzz==3.6.1
aiofiles==23.2.1
requests==2.31.0
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2